+++ vortex-array/src/array/varbin/compute/mod.rs
@@ -4,9 +4,7 @@ use vortex_scalar::Scalar;
 use crate::array::varbin::{varbin_scalar, VarBinArray};
 use crate::array::VarBinEncoding;
 use crate::compute::unary::ScalarAtFn;
-use crate::compute::{
-    ArrayCompute, ComputeVTable, FilterFn, MaybeCompareFn, Operator, SliceFn, TakeFn,
-};
+use crate::compute::{ArrayCompute, CompareFn, ComputeVTable, FilterFn, SliceFn, TakeFn};
 use crate::{ArrayDType, ArrayData};
 
 mod compare;
@@ -15,8 +13,8 @@ mod slice;
 mod take;
 
 impl ArrayCompute for VarBinArray {
-    fn compare(&self, other: &ArrayData, operator: Operator) -> Option<VortexResult<ArrayData>> {
-        MaybeCompareFn::maybe_compare(self, other, operator)
+    fn compare(&self) -> Option<&dyn CompareFn> {
+        Some(self)
     }
 }
