+++ vortex-array/src/array/varbinview/compute/mod.rs
@@ -1,33 +1,25 @@
-use std::sync::Arc;
+mod compare;
 
-use arrow_array::cast::AsArray;
-use arrow_array::types::ByteViewType;
-use arrow_array::{Datum, GenericByteViewArray};
 use arrow_buffer::ScalarBuffer;
-use arrow_ord::cmp;
-use arrow_schema::DataType;
 use itertools::Itertools;
 use num_traits::AsPrimitive;
 use vortex_buffer::Buffer;
 use vortex_dtype::{match_each_integer_ptype, PType};
-use vortex_error::{vortex_bail, VortexResult};
+use vortex_error::VortexResult;
 use vortex_scalar::Scalar;
 
 use crate::array::varbin::varbin_scalar;
 use crate::array::varbinview::{VarBinViewArray, VIEW_SIZE_BYTES};
-use crate::array::{ConstantArray, PrimitiveArray, VarBinViewEncoding};
-use crate::arrow::FromArrowArray;
+use crate::array::{PrimitiveArray, VarBinViewEncoding};
 use crate::compute::unary::ScalarAtFn;
-use crate::compute::{
-    slice, ArrayCompute, ComputeVTable, MaybeCompareFn, Operator, SliceFn, TakeFn, TakeOptions,
-};
+use crate::compute::{slice, ArrayCompute, CompareFn, ComputeVTable, SliceFn, TakeFn, TakeOptions};
 use crate::validity::Validity;
 use crate::variants::PrimitiveArrayTrait;
-use crate::{ArrayDType, ArrayData, ArrayLen, IntoArrayData, IntoArrayVariant, IntoCanonical};
+use crate::{ArrayDType, ArrayData, IntoArrayData, IntoArrayVariant};
 
 impl ArrayCompute for VarBinViewArray {
-    fn compare(&self, other: &ArrayData, operator: Operator) -> Option<VortexResult<ArrayData>> {
-        MaybeCompareFn::maybe_compare(self, other, operator)
+    fn compare(&self) -> Option<&dyn CompareFn> {
+        Some(self)
     }
 }
 
@@ -131,89 +123,12 @@ fn take_views_unchecked<I: AsPrimitive<usize>>(
     )
 }
 
-impl MaybeCompareFn for VarBinViewArray {
-    fn maybe_compare(
-        &self,
-        other: &ArrayData,
-        operator: Operator,
-    ) -> Option<VortexResult<ArrayData>> {
-        other.as_constant().map(|rhs_const| {
-            compare_constant(self, &ConstantArray::new(rhs_const, self.len()), operator)
-        })
-    }
-}
-
-fn compare_constant(
-    lhs: &VarBinViewArray,
-    rhs: &ConstantArray,
-    operator: Operator,
-) -> VortexResult<ArrayData> {
-    let arrow_lhs = lhs.clone().into_canonical()?.into_arrow()?;
-    let constant = Arc::<dyn Datum>::try_from(&rhs.owned_scalar())?;
-
-    match arrow_lhs.data_type() {
-        DataType::BinaryView => {
-            compare_constant_arrow(arrow_lhs.as_binary_view(), constant, operator)
-        }
-        DataType::Utf8View => {
-            compare_constant_arrow(arrow_lhs.as_string_view(), constant, operator)
-        }
-        _ => {
-            vortex_bail!("Cannot compare VarBinViewArray with non-binary type");
-        }
-    }
-}
-
-fn compare_constant_arrow<T: ByteViewType>(
-    lhs: &GenericByteViewArray<T>,
-    rhs: Arc<dyn Datum>,
-    operator: Operator,
-) -> VortexResult<ArrayData> {
-    let rhs = rhs.as_ref();
-    let array = match operator {
-        Operator::Eq => cmp::eq(lhs, rhs)?,
-        Operator::NotEq => cmp::neq(lhs, rhs)?,
-        Operator::Gt => cmp::gt(lhs, rhs)?,
-        Operator::Gte => cmp::gt_eq(lhs, rhs)?,
-        Operator::Lt => cmp::lt(lhs, rhs)?,
-        Operator::Lte => cmp::lt_eq(lhs, rhs)?,
-    };
-    Ok(ArrayData::from_arrow(&array, true))
-}
-
 #[cfg(test)]
 mod tests {
-    use vortex_dtype::Nullability;
-    use vortex_scalar::Scalar;
-
     use crate::accessor::ArrayAccessor;
-    use crate::array::varbinview::compute::compare_constant;
-    use crate::array::{ConstantArray, PrimitiveArray, VarBinViewArray};
-    use crate::compute::{take, Operator, TakeOptions};
-    use crate::{ArrayDType, ArrayLen, IntoArrayData, IntoArrayVariant};
-
-    #[test]
-    fn basic_test() {
-        let arr = VarBinViewArray::from_iter_nullable_str([
-            Some("one"),
-            Some("two"),
-            Some("three"),
-            Some("four"),
-            Some("five"),
-            Some("six"),
-        ]);
-
-        let s = Scalar::utf8("seven".to_string(), Nullability::Nullable);
-
-        let constant_array = ConstantArray::new(s, arr.len());
-
-        let r = compare_constant(&arr, &constant_array, Operator::Eq)
-            .unwrap()
-            .into_bool()
-            .unwrap();
-
-        assert!(r.boolean_buffer().iter().all(|v| !v));
-    }
+    use crate::array::{PrimitiveArray, VarBinViewArray};
+    use crate::compute::{take, TakeOptions};
+    use crate::{ArrayDType, IntoArrayData, IntoArrayVariant};
 
     #[test]
     fn take_nullable() {
